[svn-r13021] Cleaned up some mismatched types that were causing errors on copper in 64-bit mode.
Tested on copper.

+++ src/H5Fsuper.c
@@ -374,8 +374,8 @@ H5F_read_superblock(H5F_t *f, hid_t dxpl_id, H5G_loc_t *root_loc, haddr_t addr, 
     {
         unsigned index_flags[H5SM_MAX_NUM_INDEXES] = {0};
         unsigned minsizes[H5SM_MAX_NUM_INDEXES] = {0};
-        size_t   sohm_l2b;           /* SOHM list-to-btree cutoff    */
-        size_t   sohm_b2l;           /* SOHM btree-to-list cutoff    */
+        unsigned sohm_l2b;           /* SOHM list-to-btree cutoff    */
+        unsigned sohm_b2l;           /* SOHM btree-to-list cutoff    */
 
         HDassert(shared->sohm_nindexes > 0 && shared->sohm_nindexes <= H5SM_MAX_NUM_INDEXES);
